Implement multilayer-perceptron-wrapping model class
Again, nothing tested yet, intentionally.

Next up would be to start writing unit tests for layer, and then for the classes building off of it incrementally higher. After that would be to flesh out get methods that return copies and restricted versions of the fields that would be used by genetic algorithms.