Begin writing unit tests for Layer class
Learned about the @:access modifier, which is amazing and dramatically simplifies unit testing, as it doesn't require me to jump through hoops to access private fields and methods, write obtuse access methods, or cram unit tests into the class being tested itself. Love it.

Next up is to write a test for the remaining public method in the Layer class.